fix: clone the id chunk before pickle dump to avoid dumping the entire tensor (#44)
Dumping a chunk of a tensor (tensor.chunk) using pickle will dump the entire tensor into the file, as chunk is just a view of the original tensor. In this PR we copy the chunk (we only chunk id tensors thus the copy cost is not very large) before the pickle dump.