Skip mem-sharing the feature tensor in cpu mode  (#75)
Skip memory-sharing the feature tensor in cpu mode (unified tensor is not used).
When a feature tensor X is initialized with a memory-shared numpy array, memory-sharing tensor X will copy it to the shared memory. With this update, X will not be copied when feature is initialized with flag "with_gpu" as false.