[Build] split building of glt and glt_v6d (#85)
Split the building of graphlearn_torch and graphlearn_torch_vineyard on both the python setup.py and CMakeLists.txt sides. This is because for graphlearn_torch_vineyard, the cxx11_abi should be adjusted to the gcc environment. For graphlearn_torch, the cxx11_abi should always be set to false to match PyTorch.