[FIX] Use a lock to protect the critical path of sampler initialization in neighbor sampler (#110)
Samplers are lazily initialized in neighbor sampler. When multiple rpc sampling requests arrive concurrently, the states of sampler initialization may be messed up. E.g., a possible observed state is that the dict self._sampler is not None, but the samplers for some edge types do not exists in the dict.